Allow all elements to have an 'id' attribute.
The motivation behind this primarily is to allow kramdown to generate auto_ids for header elements, which can not only be used for link anchors, but also play a central role in kramdown-generated ToCs.

The downside, of course, is that you may not _want_ your users to use HTML tag IDs in their input.